added id and program props and added detprogrambyid
