+++ components/sections/PaymentCard/index.tsx
@@ -4,6 +4,8 @@ import { useRouter } from "next/router";
 import Button from "@/old-components/Button/Button";
 
 export type PaymentCardData = {
+  id?:string;
+  program?:number;
   title?: string;
   subtitle?: string;
   perks?: Array<any>;
@@ -15,11 +17,15 @@ export type PaymentCardData = {
   periodicity?: string;
   featured_price?: boolean;
   payment_provider_image?: StrapiImage;
+  metadata?:any;
+  config?:any;
 }
 
 const PaymentCard = (props: PaymentCardData) => {
 
   const {
+    id,
+    program,
     title,
     subtitle,
     perks = [],
@@ -30,11 +36,10 @@ const PaymentCard = (props: PaymentCardData) => {
     total_payment,
     periodicity,
     featured_price,
-    payment_provider_image } = props;
+    payment_provider_image, metadata, config} = props;
 
   const router = useRouter();
-
-
+  
   return (
     <div className={cn("keen-slider__slide gap-y-4 grid rounded-lg border border-surface-200 p-4 bg-white w-72 h-72 mobile:mx-auto", { "opacity-40": !checkout_url, "!border-primary-400": featured_price })}>
       <div>
@@ -105,7 +110,7 @@ const PaymentCard = (props: PaymentCardData) => {
             : null
         }
       </div>
-      {
+      {/* {
         checkout_url
           ? <div className="flex items-end justify-center">
             <Button
@@ -125,10 +130,38 @@ const PaymentCard = (props: PaymentCardData) => {
               }}
             />
           </div>
-          : <div className="flex items-center">
-            <p className="text-6 font-bold">Próximamente</p>
-          </div>
-      }
+          :  metadata && config ?<div className="flex items-end justify-center">
+          <Button
+            dark
+            data={{
+              type: "primary",
+              title: "Inscribirme ahora",
+              isExpand: true,
+            }}
+            onClick={() => {
+              if (checkout_url) {
+                router.push(`/checkout/${program}/${id}`);
+              } else {
+                return
+              }
+            }}
+          />
+        </div>:  <div className="flex items-center">
+            <p className="text-6 font-bold">Próximamente</p> */}
+          {/* </div> }*/}
+        <div className="flex items-end justify-center">
+        <Button
+          dark
+          data={{
+            type: "primary",
+            title: "Inscribirme ahora",
+            isExpand: true,
+          }}
+          onClick={() => {
+              router.push(`/checkout/${program}/${id}`);
+          }}
+        />
+      </div>
     </div>
   )
